feat : [YK] adding profile & edit profile features

+++ lib/data/api/auth/model/current_total_nutrition_model.dart
@@ -2,9 +2,9 @@ class CurrentTotalNutrition {
   int? totalKaloriMakan;
   int? totalKaloriAktivitas;
   int? totalKaloriSekarang;
-  double? totalLemak;
-  double? totalProtein;
-  double? totalKarbohidrat;
+  int? totalLemak;
+  int? totalProtein;
+  int? totalKarbohidrat;
 
   CurrentTotalNutrition(
       {this.totalKaloriMakan,
